GOOD NEW EVERYONE! Gerber basically said we could steal code from Malloy, so here we go. I renamed the variable in .... keep reading below
Renamed the variable for file name in make file, basically copied the
piss out of Malloy's code, only made the memory adjustment to 1024 which
we needed, and move his stuff from the header file into the main file,
because we don't need this to be any more complicated. I'm gonna start
poking at this now and try to get rid of the superflous stuff, as well
as test these functions to make sure they do what we want them to do,
but this looks like a super big help, and we probably don't have to
stress nealy as much as I though we did. High five!